api/requirements.txt
Django==2.0.5
psycopg2-binary==2.7.5
pytz==2018.4
djangorestframework==3.8.2
django-rest-swagger==2.2.0
PyYAML===3.13
django-recurrence==1.6.0
djangorestframework-jwt-custom-user==1.11.1
six==1.11.0
